% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{bytes}
\alias{bytes}
\title{Split a utf-8 encoded string into bytes}
\usage{
bytes(s)
}
\arguments{
\item{s}{a utf-8 encoded string}
}
\value{
a tibble (with extra class "tbl_bytes") with columns
- id : index of the byte
- byte: hex representation of the byte
- decimal: decimal representation of the byte
- binary: binary representation of the byte
}
\description{
Split a utf-8 encoded string into bytes
}
\examples{
bytes( "hello" )

}
