% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{runes}
\alias{runes}
\title{Split a string into unicode runes}
\usage{
runes(s)
}
\arguments{
\item{s}{a utf-8 encoded string}
}
\value{
a tibble (with extra class "tbl_runes") with columns:
- id: index of the rune in the input
- rune: name of the rune, i.e. an hex representation prefixed by "U+"
- rune_binary: binary representation of the rune
- rune_decimal: decimal representation of the rune, i.e. the index of the character in the unicode table
- utf8_bytes: hex representation of the utf8 bytes
- utf8_binary: binary representation of the utf8 bytes
- description: description of the rune, from [uni::code]
}
\description{
Split a string into unicode runes
}
\examples{
runes( "hello world" )

}
